lib/ffi/fltk1.rb: protect the Widget label from reallocation.
The Widget constructor keeps a reference to the label pointer, which
is the content of the Ruby string passed to the FFI.  Therefore we
must ensure that the string is never garbage collected or modified,
since that could free or reallocate the widget label.  We do this by
calling #dup on the value before passing it to FFI (so that we own the
only reference to it) and storing it in an instance variable (so that
it is never garbage collected).